Initial functioning build of Etch-A-Sketchy app
The app generates a square grid of 100 html elements. Currently, two
functions control the the functioning of the app. As the user moves the
cursor across the grid, each element will change color to either black
or a random color, depending on which function is active in the .js file.
Future functionality will allow user input to change grid size as well as
coloring functionality.