Add additional functionality to app
At this point, both the number of grid elements and the coloring
function applied to the grid have been made dynamic based on user
input through the use of sidebar buttons. The app begins by generating
a 50X50 grid, with the black marker function selected. The coloring
function can be changed to a random color or to white, simulating an
eraser. At this time, the triggering event for grid coloring has been
left as a mouseover event, both to simulate an original etch-a-sketch,
and to keep the code simple. In the future, this may be modified to
accept click-and-drag in order to offer better user control over the
placement of activated grid elements.